+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Version 0.1.4.0
 
 <https://github.com/mstksg/emd/releases/tag/v0.1.4.0>
 
-*   `transposeSpectrum` added to *Numeric.HHT* module, for an alternate
+*   `hhtSpetrumT` added to *Numeric.HHT* module, for an alternate
     representation of the Hilbert Spectrum.
 
 Version 0.1.3.0